HSEARCH-2269 Moved *String2FieldBridge*Adaptor classes to an impl package that is accessible to other modules
This will allow to use them in the Elasticsearch module, while clearly notifying users they should not access this package.
The target package is already exported for OSGi.
