+++ engine/src/main/java/org/hibernate/search/bridge/builtin/impl/BuiltinIterableBridge.java
@@ -11,6 +11,7 @@ import org.hibernate.search.bridge.FieldBridge;
 import org.hibernate.search.bridge.StringBridge;
 import org.hibernate.search.bridge.builtin.DefaultStringBridge;
 import org.hibernate.search.bridge.builtin.IterableBridge;
+import org.hibernate.search.bridge.util.impl.String2FieldBridgeAdaptor;
 
 /**
  * An implementation of {@link org.hibernate.search.bridge.builtin.IterableBridge} that can be used with Hibernate Search query DSL.
