+++ engine/src/main/java/org/hibernate/search/bridge/builtin/impl/BuiltinMapBridge.java
@@ -10,6 +10,7 @@ import org.hibernate.search.bridge.FieldBridge;
 import org.hibernate.search.bridge.StringBridge;
 import org.hibernate.search.bridge.builtin.DefaultStringBridge;
 import org.hibernate.search.bridge.builtin.MapBridge;
+import org.hibernate.search.bridge.util.impl.String2FieldBridgeAdaptor;
 
 /**
  * An implementation of {@link org.hibernate.search.bridge.builtin.MapBridge} that can be used with Hibernate Search query DSL.
