+++ engine/src/main/java/org/hibernate/search/bridge/util/impl/TwoWayString2FieldBridgeAdaptor.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.bridge.builtin.impl;
+package org.hibernate.search.bridge.util.impl;
 
 import org.apache.lucene.document.Document;
 import org.apache.lucene.index.IndexableField;
